# <chrono>: Cache QueryPerformanceFrequency() and divide just once