--- a/stl/inc/chrono
+++ b/stl/inc/chrono
@@ -633,9 +633,9 @@ namespace chrono {
                 const long long _Whole = (_Ctr_base / _Freq) * period::den;
                 const long long _Part  = (_Ctr_base % _Freq) * period::den / _Freq;
                 _Result_base           = _Whole + _Part;
-                if (_Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Freq, _Freq, LLONG_MAX) && 
-                    _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Ctr_base, _Ctr_base, LLONG_MAX) &&
-                    _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Result_base, _Result_base, LLONG_MAX)) {
+                if (_Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Freq, _Freq, LLONG_MAX)
+                    && _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Ctr_base, _Ctr_base, LLONG_MAX)
+                    && _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Result_base, _Result_base, LLONG_MAX)) {
                     return time_point(duration(_Result_base));
                 }
             }