--- a/stl/inc/chrono
+++ b/stl/inc/chrono
@@ -617,30 +617,28 @@ namespace chrono {
                 volatile long long _Ctr_base    = LLONG_MAX;
                 volatile long long _Result_base = LLONG_MAX;
             } _Cached;
-            long long _Freq;
-            long long _Ctr_base;
-            long long _Result_base;
-            for (;;) {
-                _Freq        = _Atomic_load_ll_relaxed(&_Cached._Freq);
-                _Ctr_base    = _Atomic_load_ll_relaxed(&_Cached._Ctr_base);
-                _Result_base = _Atomic_load_ll_relaxed(&_Cached._Result_base);
-                if (_Freq != LLONG_MAX && _Ctr_base != LLONG_MAX && _Result_base != LLONG_MAX) {
-                    break;
-                }
-                static_assert(period::num == 1, "This assumes period::num == 1.");
-                _Freq                  = _Query_perf_frequency();
-                _Ctr_base              = _Query_perf_counter();
-                const long long _Whole = (_Ctr_base / _Freq) * period::den;
-                const long long _Part  = (_Ctr_base % _Freq) * period::den / _Freq;
-                _Result_base           = _Whole + _Part;
-                if (_Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Freq, _Freq, LLONG_MAX)
-                    && _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Ctr_base, _Ctr_base, LLONG_MAX)
-                    && _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Result_base, _Result_base, LLONG_MAX)) {
-                    return time_point(duration(_Result_base));
-                }
+            long long _Freq              = _Atomic_load_ll_relaxed(&_Cached._Freq);
+            const long long _Ctr_base    = _Atomic_load_ll_relaxed(&_Cached._Ctr_base);
+            const long long _Result_base = _Atomic_load_ll_relaxed(&_Cached._Result_base);
+            if (_Freq != LLONG_MAX && _Ctr_base != LLONG_MAX && _Result_base != LLONG_MAX) {
+                // Fast path
+                const long long _Ctr = _Query_perf_counter();
+                return time_point(duration((_Ctr - _Ctr_base) * period::den / _Freq + _Result_base));
             }
+            // Caclulate with two divisions to prevent overflow
+            _Freq                = _Query_perf_frequency();
             const long long _Ctr = _Query_perf_counter();
-            return time_point(duration((_Ctr - _Ctr_base) * period::den / _Freq + _Result_base));
+            static_assert(period::num == 1, "This assumes period::num == 1.");
+            const long long _Whole  = (_Ctr / _Freq) * period::den;
+            const long long _Part   = (_Ctr % _Freq) * period::den / _Freq;
+            const long long _Result = _Whole + _Part;
+            if (_Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Freq, _Freq, LLONG_MAX)) {
+                // This is the first result, save current result as base for fast path
+                _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Ctr_base, _Ctr, LLONG_MAX);
+                _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Result_base, _Result, LLONG_MAX);
+            }
+            // if _Result is not saved as first, it is still compatible with fast result
+            return time_point(duration(_Result));            
 #else // ^^^ known hardware && !defined(_M_CEE_PURE) / unknown hardware || defined(_M_CEE_PURE) vvv
             const long long _Freq = _Query_perf_frequency(); // doesn't change after system boot
             const long long _Ctr  = _Query_perf_counter();