--- a/stl/inc/chrono
+++ b/stl/inc/chrono
@@ -609,6 +609,18 @@ namespace chrono {
         using time_point                = chrono::time_point<steady_clock>;
         static constexpr bool is_steady = true;
 
+        static long long _Scale_large_counter(const long long _Ctr, const long long _Freq) {
+            static_assert(period::num == 1, "This assumes period::num == 1.");
+            // Instead of just having "(_Ctr * period::den) / _Freq",
+            // the algorithm below prevents overflow when _Ctr is sufficiently large.
+            // It assumes that _Freq * period::den does not overflow, which is currently true for nano period.
+            // It is not realistic for _Ctr to accumulate to large values from zero with this assumption,
+            // but the initial value of _Ctr could be large.
+            const long long _Whole = (_Ctr / _Freq) * period::den;
+            const long long _Part  = (_Ctr % _Freq) * period::den / _Freq;
+            return _Whole + _Part;
+        }
+
         _NODISCARD static time_point now() noexcept { // get current time
 #if (defined(_M_IX86) || defined(_M_X64) || defined(_M_ARM) || defined(_M_ARM64)) && !defined(_M_CEE_PURE)
             // Implement atomics avoiding <atomic> header dependency
@@ -628,10 +640,7 @@ namespace chrono {
             // Caclulate with two divisions to prevent overflow
             _Freq                = _Query_perf_frequency();
             const long long _Ctr = _Query_perf_counter();
-            static_assert(period::num == 1, "This assumes period::num == 1.");
-            const long long _Whole  = (_Ctr / _Freq) * period::den;
-            const long long _Part   = (_Ctr % _Freq) * period::den / _Freq;
-            const long long _Result = _Whole + _Part;
+            const long long _Result = _Scale_large_counter(_Ctr, _Freq);
             if (_Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Freq, _Freq, LLONG_MAX)) {
                 // This is the first result, save current result as base for fast path
                 _Atomic_compare_exchange_strong_ll_cst_seq(&_Cached._Ctr_base, _Ctr, LLONG_MAX);
@@ -642,15 +651,7 @@ namespace chrono {
 #else // ^^^ known hardware && !defined(_M_CEE_PURE) / unknown hardware || defined(_M_CEE_PURE) vvv
             const long long _Freq = _Query_perf_frequency(); // doesn't change after system boot
             const long long _Ctr  = _Query_perf_counter();
-            static_assert(period::num == 1, "This assumes period::num == 1.");
-            // Instead of just having "(_Ctr * period::den) / _Freq",
-            // the algorithm below prevents overflow when _Ctr is sufficiently large.
-            // It assumes that _Freq * period::den does not overflow, which is currently true for nano period.
-            // It is not realistic for _Ctr to accumulate to large values from zero with this assumption,
-            // but the initial value of _Ctr could be large.
-            const long long _Whole = (_Ctr / _Freq) * period::den;
-            const long long _Part  = (_Ctr % _Freq) * period::den / _Freq;
-            return time_point(duration(_Whole + _Part));
+            return time_point(duration(_Scale_large_counter(_Ctr, _Freq)));
 #endif // (defined(_M_IX86) || defined(_M_X64) || defined(_M_ARM) || defined(_M_ARM64)) && !defined(_M_CEE_PURE)
         }