--- a/stl/inc/chrono
+++ b/stl/inc/chrono
@@ -8,7 +8,6 @@
 #define _CHRONO_
 #include <yvals_core.h>
 #if _STL_COMPILER_PREPROCESSOR
-#include <intrin0.h>
 #include <limits>
 #include <ratio>
 #include <time.h>