… stl/inc/xatomic.h