@AlexGuteniev 2020-03-29
Benchmark has shown noticeable improvement on x86 and x64. 
I don't have access to ARM or ARM64 to benchmark.

Benchmark is 50'000'000 reading of `Clock::now()` and storing them into volatile variable.
(Without volatile variable division is optimized away)

Benchmark results:

| implementation|x86|x64|
|------------------|-----|---|
|steady_clock after #646 (#646 does not help much on my system)|1940 ms|3113 ms|
|after proposed change|1308 ms|2334 ms|
|after proposed change with magic static instead of atomic|1282 ms|2307 ms|

Benchmark
```
#include <iostream>
#include <chrono>
#include <atomic>

template<class Derived>
struct clock_base
{
    using  rep = std::int64_t;
    using  period = typename std::chrono::steady_clock::period;
    using  duration = std::chrono::duration<rep, period>;
    using  time_point = std::chrono::time_point<Derived, duration>;

    static constexpr bool is_steady = true;
};

struct magic_static_clock final : clock_base<magic_static_clock>
{
    static time_point now()
    {
        static_assert(period::num == 1, "Assuming even second fraction");
        struct clock_params
        {
            const rep freq = _Query_perf_frequency();
            const rep quot = period::den / freq;
            const rep rem = period::den % freq;
        };
        static const clock_params params;
        const rep counter = _Query_perf_counter();
        return time_point(duration(
            params.quot * counter + params.rem * counter / params.freq));
    }
};

struct atomic_cache_clock final : clock_base<atomic_cache_clock>
{
    static time_point now()
    {
        static_assert(period::num == 1, "Assuming even second fraction");
        static struct
        {
            std::atomic<rep> freq = -1;
            std::atomic<rep> quot = -1;
            std::atomic<rep> rem = -1;
        } cache;
        rep freq = cache.freq.load(std::memory_order_relaxed);
        rep quot = cache.quot.load(std::memory_order_relaxed);
        rep rem = cache.rem.load(std::memory_order_relaxed);
        if (freq == -1 || quot == -1 || rem == -1)
        {
            freq = _Query_perf_frequency();
            quot = period::den / freq;
            rem = period::den % freq;
            cache.freq.store(freq, std::memory_order_relaxed);
            cache.quot.store(quot, std::memory_order_relaxed);
            cache.rem.store(rem, std::memory_order_relaxed);
        }
        const rep counter = _Query_perf_counter();
        return time_point(duration(quot * counter + rem * counter / freq));
    }
};

constexpr auto N = 50'000'000;

volatile std::int64_t now_result;

template<class Clock>
void test_clock(const char* name)
{
    auto t = std::chrono::steady_clock::now();
    for (std::size_t i = 0; i < N; i++) {
        now_result = Clock::now().time_since_epoch().count();
    }
    auto d = std::chrono::steady_clock::now() - t;
    auto ms = std::chrono::duration_cast<std::chrono::duration<double, std::milli>>(d);
    std::cout << ms.count() << "ms " << name << "\n";
}

int main()
{
    test_clock<std::chrono::steady_clock>("original");
    test_clock<atomic_cache_clock>("atomic cache");
    test_clock<magic_static_clock>("magic static");
}
```