@AlexGuteniev 2020-03-29
I discovered the optimization is incorrect. Original code:
```
const long long _Part  = (_Ctr % _Freq) * period::den / _Freq;
```
assumes that _Freq*period::den fits 64 bit.

after change the code assumes _Freq*_Ctr fits 64 bit, which is what the original approach protects against.