@StephanTLavavej 2020-04-06
This is paying a significant amount of implementation complexity - I think we need a strong performance justification for making this change. Can you provide performance numbers for `steady_clock::now`:
1. Before #646
2. After #646, but before this PR
3. After this PR