@AlexGuteniev 2020-04-07
@StephanTLavavej ,
See https://github.com/AlexGuteniev/benchmark_653

My results are 
6042.98  before 646
5904.46  after 646
4520.97  after 653

It would be simpler if it would go as vNext issue, then no fake atomic needed, it would be possible to move the most of it to .cpp.