@AlexGuteniev 2020-04-07
Another option to achieve the same is to use `_mul128` / `_div128`.
But then the optimization will be applied to x64 only, and it will take bringing in  `_mul128` / `_div128` intrinsics, not currently available in `"intrin0.h"`