@AlexGuteniev 2020-04-07
(There's also an option which you you would definitely reject, but I'm considering for my own use, it is using https://github.com/ridiculousfish/libdivide for division)