@cbezault 2020-04-09
Here are the benchmark results on an ARM64 machine.
```
5983.27  before 646
5791.83  after 646
5269.29  after 653