@AlexGuteniev 2020-04-11
I've incorporated #658 , as it is merged.

A couple of more observations: 

- I think it is needed to adjust result by adding `_Result_base` back even in vNext, even if we do just call_once on initial initialization. One reason is that `time_point` could be shared between (statically linked) library instances or even between processes. Another is that avoiding `call_once` may still be preferred due to being lock-free.
- Such adjustment, when `_Ctr_base` is subtracted and `_Result_base` is added result in precision loss. This precision loss is even in worst case less than the divisor, so less than nanosecond, and happens only before every thread observe cache initialization or between STL instances, so it is fine.