@cbezault 2020-05-08
I agree with Alex's assessment that at worse we lose precision that is strictly less than one nanosecond. This can only occur when a thread both does not observe a fully populated cache and when it is not the thread which ultimately populates the cache.

The other case of two processes populating the cache with different `_Ctr_base`s and the same `_Result_base` causing a subsequent call of `now()` where both processes get the same result from `_Query_perf_counter()` giving different `time_point`s also seems acceptable. Again the difference will be at most one nanosecond.