@StephanTLavavej 2020-05-14
Thanks for this significant performance improvement! I would have never noticed that this was possible, so I'm glad you did. 😺