@jpark37 2020-11-14
I see this has been backed out since, but is there a reason to use atomics and cache to function-local static variables? Wouldn't it be simpler and faster (no branch) to pull those variables out of the function and initialize them before main()? I see _Query_perf_frequency() in xtime.cpp is implemented similarly.