@BillyONeal 2020-11-15
@jpark That would add a dynamic initializer to the STL and I believe we don't have any of those right now. I seem to recall adding any causing issues with `/clr:pure` code that tried to statically link the STL in a partially trusted context, since they can't necessarily run any native code here.

But my memory is a bit foggy on the specifics. I know we had to avoid the dynamic initializer but I don't remember the exact technical limitations forcing it.