@BillyONeal 2020-11-15
(Also I don't know if `QueryPerformanceFrequency` is in the "safe to be called from DllMain" subset since the default answer for "is this safe to call from DllMain" is "no".)