@jpark37 2020-11-15
Thanks for the info.

Aren't people allowed to use STL before main() in general? I've seen people declare STL collections as global variables at least. Right now, if someone calls steady_clock::now() in a global variable initializer, and it's the first call into _Query_perf_frequency(), that will call QueryPerformanceFrequency().

Thinking about it more, it may be unsafe to call my version of _Query_perf_frequency() from another initializer before main() because of not being able to control the order of initialization across translation units, so freq_cached might still be 0 with unlucky order. I don't know if the story is any different with /MT vs. /MD.